promocao

+++ api_promo/models/promocao_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Column, Date, ForeignKey
+from sqlalchemy import Integer, String, Column, Date, ForeignKey,BigInteger
 from sqlalchemy.orm import relationship
 
 from core.configs import settings
@@ -13,7 +13,7 @@ class PromocaoModel(settings.DBBaseModel):
     data_validade = Column(String(256), nullable=True)
     usuario_id = Column(Integer, ForeignKey('usuario.id'))
     estabelecimento_id = Column(Integer, ForeignKey('estabelecimento.id'))
-    produto_id = Column(Integer, ForeignKey('produto.id'))
+    produto_id = Column(BigInteger, ForeignKey('produto.id'))
     criador = relationship("UsuarioModel", back_populates='promocoes', lazy='joined')
     produto = relationship("ProdutoModel", back_populates="promocoes", lazy='joined')
     estabelecimento = relationship("EstabelecimentoModel", back_populates="promocoes", lazy='joined')
